Episodes

Each episodes' statistics. This dataset is made by collapsing challenges by series and episode.

Usage

episodes

Format

A data frame with 94 rows representing individual episodes per series and 10 variables:

series

An integer denoting UK series (1-10).

episode

An integer denoting episode number within a series.

bakers_appeared

an integer denoting the number of bakers who appeared in a given episode.

bakers_out

An integer denoting the number of bakers were either eliminated, left at will, or left due to illness in that episode.

bakers_remaining

An integer denoting the number of bakers who remained at the end of a given episode.

star_bakers

An integer denoting the number of bakers crowned Star Baker in a given episode.

technical_winners

An integer denoting the number of bakers who placed first in the technical challenge.

sb_name

A character string with the names of the bakers who won Star Baker in a given episode, separated by commas.

winner_name

A character string with the name of the series winner. Only present for final episodes within each series.

eliminated

A character string with the name of the bakers who left a given episode, separated by commas.

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off#Format

Examples

if (require('tibble')) {
  episodes
 }
#> # A tibble: 94 × 10
#>    series episode bakers_appeared bakers_out bakers_remaining star_bakers
#>     <dbl>   <dbl>           <int>      <int>            <int>       <int>
#>  1      1       1              10          2                8           0
#>  2      1       2               8          2                6           0
#>  3      1       3               6          1                5           0
#>  4      1       4               5          1                4           0
#>  5      1       5               4          1                3           0
#>  6      1       6               3          0                3           0
#>  7      2       1              12          1               11           1
#>  8      2       2              11          1               10           1
#>  9      2       3              10          2                8           1
#> 10      2       4               8          1                7           2
#> # … with 84 more rows, and 4 more variables: technical_winners <int>,
#> #   sb_name <chr>, winner_name <chr>, eliminated <chr>
head(episodes)
#> # A tibble: 6 × 10
#>   series episode bakers_appeared bakers_out bakers_remaining star_bakers
#>    <dbl>   <dbl>           <int>      <int>            <int>       <int>
#> 1      1       1              10          2                8           0
#> 2      1       2               8          2                6           0
#> 3      1       3               6          1                5           0
#> 4      1       4               5          1                4           0
#> 5      1       5               4          1                3           0
#> 6      1       6               3          0                3           0
#> # … with 4 more variables: technical_winners <int>, sb_name <chr>,
#> #   winner_name <chr>, eliminated <chr>