Ratings

Each episode's ratings plus original airdates in the UK (by series) and US (by season). This dataset is made by joining the ratings_raw and seasons_raw datasets by both series and episode.

Usage

ratings

Format

A data frame with 74 rows representing individual episodes and 10 variables:

series

An integer denoting UK series (1-8).

episode

An integer denoting episode number within a series.

uk_airdate

A date denoting original airdate of episode in the UK.

viewers_7day

The number of viewers in millions within a 7-day window from airdate.

viewers_28day

The number of viewers in millions within a 28-day window from airdate.

network_rank

An integer denoting the episode's weekly ranking within network (note that the networks changed for series 8 and later).

channels_rank

An integer denoting the episode's weekly ranking across all channels.

bbc_iplayer_requests

Number of BBC iPlayer requests (note: not available for all series, and only for series that aired on the BBC).

us_season

An integer denoting US season (1-5).

us_airdate

A date denoting original airdate of episode in the US, according to pbs.org.

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off_(series_1)#Ratings

Examples

if (require('tibble')) {
  ratings
 }
#> # A tibble: 74 × 10
#>    series episode uk_airdate viewers_7day viewers_28day network_rank
#>     <int>   <int> <date>            <dbl>         <dbl>        <int>
#>  1      1       1 2010-08-17         2.24             7           NA
#>  2      1       2 2010-08-24         3                3           NA
#>  3      1       3 2010-08-31         3                2           NA
#>  4      1       4 2010-09-07         2.6              4           NA
#>  5      1       5 2010-09-14         3.03             1           NA
#>  6      1       6 2010-09-21         2.75             1           NA
#>  7      2       1 2011-08-16         3.1              2           NA
#>  8      2       2 2011-08-23         3.53             2           NA
#>  9      2       3 2011-08-30         3.82             1           NA
#> 10      2       4 2011-09-06         3.6              1           NA
#> # … with 64 more rows, and 4 more variables: channels_rank <int>,
#> #   bbc_iplayer_requests <dbl>, us_season <int>, us_airdate <date>