Data about each series aired in the UK (raw)

This dataset has one row per series aired in the UK as 'The Great British Bake Off'. When did a series premiere in the UK? When was the finale episode? How many viewers on average did a series have?

Usage

series_raw

Format

A data frame with 10 rows representing individual series and 11 variables:

series

A factor denoting UK series (1-10).

episodes

An integer denoting total number of episodes within series.

premiere

A date denoting original airdate of series premiere episode in the UK.

finale

A date denoting original airdate of series finale episode in the UK.

winner

A character string denoting name of series winner.

avg_uk_viewers

Average UK viewers in millions. Also see ratings.

day_of_week

A character string denoting weekday that series' episodes aired in UK.

timeslot

time of day that series' episodes aired in UK (24-hour clock).

channel

A character string denoting channel that series aired on: one of BBC One, BBC Two, or Channel 4.

runner_up_1

A character string denoting the name of one of the bakers who was a series runner-up. The number is not meaningful.

runner_up_2

A character string denoting the name of one of the bakers who was a series runner-up. The number is not meaningful.

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off#Series_overview

Examples

if (require('tibble')) {
  series_raw
 }
#> # A tibble: 10 × 11
#>    series episodes premiere   finale     winner  avg_u…¹ day_o…² times…³ channel
#>    <fct>     <int> <date>     <date>     <chr>     <dbl> <chr>   <chr>   <chr>  
#>  1 1             6 2010-08-17 2010-09-21 Edd Ki…    2.77 Tuesday 8:00pm  BBC Two
#>  2 2             8 2011-08-14 2011-10-04 Joanne…    4    Tuesday 8:00pm  BBC Two
#>  3 3            10 2012-08-14 2012-10-16 John W…    5    Tuesday 8:00pm  BBC Two
#>  4 4            10 2013-08-20 2013-10-22 France…    7.35 Tuesday 8:00pm  BBC Two
#>  5 5            10 2014-08-06 2014-10-08 Nancy …   10.0  Wednes… 8:00pm  BBC One
#>  6 6            10 2015-08-05 2015-10-07 Nadiya…   12.5  Wednes… 8:00pm  BBC One
#>  7 7            10 2016-08-24 2016-10-26 Candic…   13.8  Wednes… 8:00pm  BBC One
#>  8 8            10 2017-08-29 2017-10-31 Sophie…    9.29 Tuesday 8:00pm  Channe…
#>  9 9            10 2018-08-28 2018-10-30 Rahul …    9.3  Tuesday 8:00pm  Channe…
#> 10 10           10 2019-08-27 2019-10-29 David …    9.24 Tuesday 8:00pm  Channe…
#> # … with 2 more variables: runner_up_1 <chr>, runner_up_2 <chr>, and
#> #   abbreviated variable names ¹​avg_uk_viewers, ²​day_of_week, ³​timeslot
head(series_raw)
#> # A tibble: 6 × 11
#>   series episodes premiere   finale     winner   avg_u…¹ day_o…² times…³ channel
#>   <fct>     <int> <date>     <date>     <chr>      <dbl> <chr>   <chr>   <chr>  
#> 1 1             6 2010-08-17 2010-09-21 Edd Kim…    2.77 Tuesday 8:00pm  BBC Two
#> 2 2             8 2011-08-14 2011-10-04 Joanne …    4    Tuesday 8:00pm  BBC Two
#> 3 3            10 2012-08-14 2012-10-16 John Wh…    5    Tuesday 8:00pm  BBC Two
#> 4 4            10 2013-08-20 2013-10-22 Frances…    7.35 Tuesday 8:00pm  BBC Two
#> 5 5            10 2014-08-06 2014-10-08 Nancy B…   10.0  Wednes… 8:00pm  BBC One
#> 6 6            10 2015-08-05 2015-10-07 Nadiya …   12.5  Wednes… 8:00pm  BBC One
#> # … with 2 more variables: runner_up_1 <chr>, runner_up_2 <chr>, and
#> #   abbreviated variable names ¹​avg_uk_viewers, ²​day_of_week, ³​timeslot