spice_test_wide {bakeoff}	R Documentation

Spice Test

Description

Results of a knowledge test from Junior Bake Off (Series 2, Episode 14).

Usage

spice_test_wide

Format

A data frame with 4 rows representing baker results from the spice-based knowledge test and 7 variables:

baker

A character string with a given name or nickname.

guess_1,guess_2,guess_3

The different guesses (in order) for what the mystery spice was.

correct_1,correct_2,correct_3

An integer whether the guess was correct (1) or wrong (0).

Examples

if (require('tibble')) {
  spice_test_wide
 }
head(spice_test_wide)

[Package bakeoff version 0.2.0 Index]