bakeoff: Data Behind the Tent at 'The Great British Bake Off'

An R package that provides access to data from 'The Great British Bake Off', as available on Wikipedia https://en.wikipedia.org/wiki/The_Great_British_Bake_Off. Contribute to this package at https://github.com/apreshill/bakeoff.

Examples

# Example usage:
library(bakeoff)
head(bakers)

#> # A tibble: 6 x 8
#>   series baker_full   baker    age occupation   hometown  baker_last baker_first
#>   <fct>  <chr>        <chr>  <dbl> <chr>        <chr>     <chr>      <chr>      
#> 1 1      "Annetha Mi… Annet…    30 Midwife      Essex     Mills      Annetha    
#> 2 1      "David Cham… David     31 Entrepreneur Milton K… Chambers   David      
#> 3 1      "Edward \"E… Edd       24 Debt collec… Bradford  Kimber     Edward     
#> 4 1      "Jasminder … Jasmi…    45 Assistant C… Birmingh… Randhawa   Jasminder  
#> 5 1      "Jonathan S… Jonat…    25 Research An… St Albans Shepherd   Jonathan   
#> 6 1      "Lea Harris" Lea       51 Retired      Midlothi… Harris     Lea        

# All information about any data set can be found in the help file:
?bakers

#> Rendering development documentation for 'bakers'

# To view a list of all data sets:
data(package = "bakeoff")

# To view a detailed list of all data sets:
vignette("bakeoff", package = "bakeoff")

#> Warning: vignette ‘bakeoff’ not found

# Some data sets include vignettes with an example analysis:
vignette("bakers", package = "bakeoff")

#> Warning: vignette ‘bakers’ not found

# To browse all vignettes:
browseVignettes(package = "bakeoff")

#> No vignettes found by browseVignettes(package = "bakeoff")