bakeoff: Data Behind the Tent at 'The Great British Bake Off'

An R package that provides access to data from 'The Great British Bake Off', as available on Wikipedia https://en.wikipedia.org/wiki/The_Great_British_Bake_Off. Contribute to this package at https://github.com/apreshill/bakeoff.

Examples

# Example usage:
library(bakeoff)
head(bakers)

#> # A tibble: 6 x 5
#>   series baker_full             age occupation                  hometown        
#>    <dbl> <chr>                <dbl> <chr>                       <chr>           
#> 1      1 Annetha Mills           30 Single mother               Essex           
#> 2      1 David Chambers          31 Entrepreneur                Milton Keynes   
#> 3      1 "Edward \"Edd\" Kim…    24 Debt collector for Yorkshi… Bradford        
#> 4      1 Jasminder Randhawa      45 Assistant Credit Control M… Birmingham      
#> 5      1 Jonathan Shepherd       25 Research Analyst            St Albans       
#> 6      1 Lea Harris              51 Retired                     Midlothian, Sco…

# All information about any data set can be found in the help file:
?bakers

#> Rendering development documentation for 'bakers'

# To view a list of all data sets:
data(package = "bakeoff")

# To view a detailed list of all data sets:
vignette("bakeoff", package = "bakeoff")

#> Warning: vignette ‘bakeoff’ not found

# Some data sets include vignettes with an example analysis:
vignette("bakers", package = "bakeoff")

#> Warning: vignette ‘bakers’ not found

# To browse all vignettes:
browseVignettes(package = "bakeoff")

#> No vignettes found by browseVignettes(package = "bakeoff")