An R package that provides access to the code and data about "The Great British Bake-Off" series, episodes, and bakers that is publicly available on Wikipedia https://en.wikipedia.org/wiki/The_Great_British_Bake_Off. Contribute to this package at https://github.com/apreshill/bakeoff.

An R package that provides access to the code and data about "The Great British Bake-Off" series, episodes, and bakers that is publicly available on Wikipedia https://en.wikipedia.org/wiki/The_Great_British_Bake_Off. Contribute to this package at https://github.com/apreshill/bakeoff.

Examples

# Example usage:
library(bakeoff)
head(bakers)

#> Error in head(bakers): object 'bakers' not found

# All information about any data set can be found in the help file:
?bakers

#> No documentation for ‘bakers’ in specified packages and libraries:
#> you could try ‘??bakers’

# To view a list of all data sets:
data(package = "bakeoff")

# To view a detailed list of all data sets:
vignette("bakeoff", package = "bakeoff")

#> Warning: vignette ‘bakeoff’ not found

# Some data sets include vignettes with an example analysis:
vignette("bakers", package = "bakeoff")

#> Warning: vignette ‘bakers’ not found

# To browse all vignettes:
browseVignettes(package = "bakeoff")

#> No vignettes found by browseVignettes(package = "bakeoff")