bakeoff_colors {bakeoff}	R Documentation

Extract named bakeoff colors as hex codes

Description

Get a vector of hexadecimal color codes, or, extract a subset of colors as a using defined color names (available using names(bakeoff_colors())).

Usage

bakeoff_colors(...)

Arguments

...
Character names of bakeoff colors. If none are specified, returns all. If any are specified, they can be listed by names in quotes (no need to combine with c()).

Value

A named character vector of hex colors.

Examples

bakeoff_colors()
bakeoff_colors("riptide")
bakeoff_colors("baltic", "yellow")
names(bakeoff_colors())

if (require('scales')) {
  scales::show_col(bakeoff_colors(), label = FALSE)
}

[Package bakeoff version 0.2.0 Index]