Each episodes' statistics

This tibble is made by collapsing the challenge_results tibbles by series and episode.

episode_results

Format

A data frame with 94 rows representing individual episodes per series and 10 variables:

series

an integer denoting UK series (1-10)

episode

an integer denoting episode number within a series

bakers_appeared

an integer denoting the number of bakers who appeared in a given episode

bakers_out

an integer denoting the number of bakers were either eliminated, left at will, or left due to illness in that episode

bakers_remaining

an integer denoting the number of bakers who remained at the end of a given episode

star_bakers

an integer denoting the number of bakers crowned Star Baker in a given episode

technical_winners

an integer denoting the number of bakers who placed first in the technical challenge

sb_name

a character string with the names of the bakers who won Star Baker in a given episode, separated by commas

winner_name

a character string with the name of the series winner- only present for final episodes within each series

eliminated

a character string with the name of the bakers who left a given episode, separated by commas

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off#Format