ratings {bakeoff}	R Documentation

Ratings

Description

This dataset has one row per episode and includes ratings plus original airdates in the UK (by series) and US (by season). This dataset is made by joining the ratings_raw and seasons_raw datasets by both series and episode.

Usage

ratings

Format

A data frame with 94 rows representing individual episodes and 11 variables:

series

An integer denoting UK series (1-10).

episode

An integer denoting episode number within a series.

uk_airdate

A date denoting original airdate of episode in the UK.

viewers_7day

The number of viewers in millions within a 7-day window from airdate.

viewers_28day

The number of viewers in millions within a 28-day window from airdate.

network_rank

An integer denoting the episode's weekly ranking within network (note that the networks changed for series 8 and later).

channels_rank

An integer denoting the episode's weekly ranking across all channels.

bbc_iplayer_requests

Number of BBC iPlayer requests (note: not available for all series, and only for series that aired on the BBC).

us_season

An integer denoting US season (1-5).

us_airdate

A date denoting original airdate of episode in the US, according to pbs.org.

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off_(series_1)#Ratings

Examples

if (require('tibble')) {
  ratings
 }
head(ratings)

[Package bakeoff version 0.2.0 Index]