Each episode's ratings plus original airdates in the UK (by series) and US (by season)

This joined tibble is made by joining the ratings and seasons simple tibbles by both series and episode.

ratings_seasons

Format

A data frame with 74 rows representing individual episodes and 10 variables:

series

an integer denoting UK series (1-8)

episode

an integer denoting episode number within a series

uk_airdate

a date denoting original airdate of episode in the UK

viewers_7day

number of viewers in millions within a 7-day window from airdate

viewers_28day

number of viewers in millions within a 28-day window from airdate

network_rank

an integer denoting the episode's weekly ranking within network (note that the networks changed for series 8 and later)

channels_rank

an integer denoting the episode's weekly ranking across all channels

bbc_iplayer_requests

number of BBC iPlayer requests (note: not available for all series, and only for series that aired on the BBC)

us_season

an integer denoting US season (1-5)

us_airdate

a date denoting original airdate of episode in the US, according to pbs.org

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off_(series_1)#Ratings