Overview of each series aired in the UK

Metadata for all 10 GBBO series

series

Format

A data frame with 10 rows representing individual series and 11 variables:

series

an integer denoting UK series (1-10)

episodes

an integer denoting total number of episodes within series

premiere

a date denoting original airdate of series premiere episode in the UK

finale

a date denoting original airdate of series finale episode in the UK

winner

a character string denoting name of series winner

avg_uk_viewers

average UK viewers in millions; also see ratings

day_of_week

a character string denoting weekday that series' episodes aired in UK

timeslot

time of day that series' episodes aired in UK (24-hour clock)

channel

a character string denoting channel that series aired on: one of BBC One, BBC Two, or Channel 4

runner_up_1

a character string denoting the name of one of the bakers who was a series runner-up; number is not meaningful

runner_up_2

a character string denoting the name of one of the bakers who was a series runner-up; number is not meaningful

Source

See https://en.wikipedia.org/wiki/The_Great_British_Bake_Off#Series_overview